error check

+++ controllers/goals.js
@@ -1,15 +1,33 @@
 const express = require("express");
+const jwt = require("jsonwebtoken");
 const router = express.Router();
 const bodyParser = require("body-parser");
 const jsonParser = bodyParser.json();
 const goalsDao = require("../dao/goals");
 
 const jwtUtil = require("../jwt/jwtVerify");
+router.use(jwtUtil.checkToken);
 
-router.get("/goals", jwtUtil.checkToken, jsonParser, async (req, res) => {
+router.get("/goals", async (req, res) => {
   try {
     res.json({
-      data: await goalsDao.getAllGoalsOfUser(
+      data: await goalsDao.getGoalsOfSelfUser(
+        payload.id,
+        req.query.month,
+        payload.role
+      ),
+    });
+  } catch (err) {
+    res.json({
+      error: err.toString(),
+    });
+  }
+});
+
+router.get("/goals-other", jwtUtil.isAdmin, async (req, res) => {
+  try {
+    res.json({
+      data: await goalsDao.getGoalsOfSelfUser(
         req.query.id,
         req.query.month,
         req.query.role
@@ -22,13 +40,24 @@ router.get("/goals", jwtUtil.checkToken, jsonParser, async (req, res) => {
   }
 });
 
-router.post("/employee-goals", jwtUtil.checkToken, jsonParser, async (req, res) => {
+router.post("/employee-goals", jsonParser, async (req, res) => {
+  const payload = jwt.verify(
+    req.headers.authorization.split(" ")[1],
+    "SecretKey"
+  );
+  const userId = parseInt(req.query.user_id);
   try {
-    const newGoals = await goalsDao.addGoalsofEmployee(req.body);
-    res.json({
-      message: "goals inserted successfully",
-      data: newGoals,
-    });
+    if (payload.id === userId) {
+      const newGoals = await goalsDao.addGoalsofEmployee(req.body);
+      res.json({
+        message: "goals inserted successfully",
+        data: newGoals,
+      });
+    } else {
+      res.json({
+        message: "you can't add to others goals",
+      });
+    }
   } catch (err) {
     res.json({
       error: err.toString(),
@@ -36,13 +65,24 @@ router.post("/employee-goals", jwtUtil.checkToken, jsonParser, async (req, res) 
   }
 });
 
-router.post("/admin-goals", jwtUtil.checkToken, jsonParser, async (req, res) => {
+router.post("/admin-goals", jwtUtil.isAdmin, jsonParser, async (req, res) => {
+  const payload = jwt.verify(
+    req.headers.authorization.split(" ")[1],
+    "SecretKey"
+  );
+  const userId = parseInt(req.query.user_id);
   try {
-    const newGoals = await goalsDao.addGoalsofAdmin(req.body);
-    res.json({
-      message: "goals inserted successfully",
-      data: newGoals,
-    });
+    if (payload.id === userId) {
+      const newGoals = await goalsDao.addGoalsofAdmin(req.body);
+      res.json({
+        message: "goals inserted successfully",
+        data: newGoals,
+      });
+    } else {
+      res.json({
+        message: "you can't add to others goals",
+      });
+    }
   } catch (err) {
     res.json({
       error: err.toString(),
@@ -50,13 +90,49 @@ router.post("/admin-goals", jwtUtil.checkToken, jsonParser, async (req, res) => 
   }
 });
 
-router.post("/super_admin-goals", jwtUtil.checkToken, jsonParser, async (req, res) => {
+router.post("/super_admin-goals", jwtUtil.isSuperAdmin, jsonParser, async (req, res) => {
+    const payload = jwt.verify(
+      req.headers.authorization.split(" ")[1],
+      "SecretKey"
+    );
+    const userId = parseInt(req.query.user_id);
+    try {
+      if (payload.id === userId) {
+        const newGoals = await goalsDao.addGoalsofSuperAdmin(req.body);
+        res.json({
+          message: "goals inserted successfully",
+          data: newGoals,
+        });
+      } else {
+        res.json({
+          message: "you can't add to others goals",
+        });
+      }
+    } catch (err) {
+      res.json({
+        error: err.toString(),
+      });
+    }
+  }
+);
+
+router.put("/update", jsonParser, async (req, res) => {
+  const payload = jwt.verify(
+    req.headers.authorization.split(" ")[1],
+    "SecretKey"
+  );
+  const userId = parseInt(req.query.user_id);
   try {
-    const newGoals = await goalsDao.addGoalsofSuperAdmin(req.body);
-    res.json({
-      message: "goals inserted successfully",
-      data: newGoals,
-    });
+    if (payload.id === userId) {
+      await goalsDao.updateGoal(req.body);
+      res.json({
+        message: "goals updated successfully",
+      });
+    } else {
+      res.json({
+        message: "you can't update others goals",
+      });
+    }
   } catch (err) {
     res.json({
       error: err.toString(),
@@ -64,25 +140,23 @@ router.post("/super_admin-goals", jwtUtil.checkToken, jsonParser, async (req, re
   }
 });
 
-router.put("/update", jwtUtil.checkToken, jsonParser,async (req, res) => {
+router.delete("/delete", async (req, res) => {
+  const payload = jwt.verify(
+    req.headers.authorization.split(" ")[1],
+    "SecretKey"
+  );
+  const userId = parseInt(req.query.user_id);
   try {
-    await goalsDao.updateGoal(req.body);
-    res.json({
-      message: "goals updated successfully",
-    });
-  } catch (err) {
-    res.json({
-      error: err.toString(),
-    });
-  }
-});
-
-router.delete("/delete", jwtUtil.checkToken, async (req, res) => {
-  try {
-    await goalsDao.deleteGoal(req.query.id);
-    res.json({
-      message: "goal deleted successfully",
-    });
+    if (payload.id === userId) {
+      await goalsDao.deleteGoal(req.query.id);
+      res.json({
+        message: "goal deleted successfully",
+      });
+    } else {
+      res.json({
+        message: "you can't delete others goals",
+      });
+    }
   } catch (err) {
     res.json({
       error: err.toString(),
